What did I do? Well, I'll tell you... I reformatted some things, optimized imports I changed a few things regarding the scala build.sbt I added the ability to be allowed to have a specific amount of homes via permission node

+++ src/main/java/world/ultravanilla/ChatEvent.scala
@@ -3,13 +3,13 @@ package world.ultravanilla
 import java.util.UUID
 
 case class ChatEvent(
-    channel: Integer,
-    sender: UUID = UUID.fromString("00000000-0000-0000-0000-000000000000"),
-    message: String = "",
-    senderNick: String = "",
-    donator: Boolean = false,
-    staff: Boolean = false,
-    isDeath: Boolean = false,
-    isAchievement: Boolean = false,
-    source: ChatSource.Value
-) {}
+                        channel: Long,
+                        sender: UUID = UUID.fromString("00000000-0000-0000-0000-000000000000"),
+                        message: String = "",
+                        senderNick: String = "",
+                        donator: Boolean = false,
+                        staff: Boolean = false,
+                        isDeath: Boolean = false,
+                        isAchievement: Boolean = false,
+                        source: ChatSource.Value
+                    ) {}
